V1.5
Added a lot of static_casts for precision

+++ 2DGameEngine/Functions.cpp
@@ -1,5 +1,6 @@
 
 #include <cmath>
+#include <string>
 #include <fstream>
 #include "Components.h"
 
@@ -66,13 +67,14 @@ float dy(const SDL_Rect& rect1, const SDL_Rect& rect2) {
     return (rect1.y + (rect1.h / static_cast<float>(2))) - (rect2.y + (rect2.h / static_cast<float>(2)));  
 }
 float distance(SDL_Rect rect1, SDL_Rect rect2)
+
 {
     float x1 = static_cast<float>(rect1.x) + static_cast<float>( rect1.w )/ 2;
     float y1 = static_cast<float>(rect1.y) + static_cast<float>(rect1.h) / 2;
     float x2 = static_cast<float>(rect2.x) + static_cast<float>(rect2.w )/ 2;
     float y2 = static_cast<float>(rect2.y) + static_cast<float>(rect2.h )/ 2;
     float delta;
-    delta = sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2));
+    delta = static_cast<float>(sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2)));
     return delta;
 }
